+++ source/main/providers/config-provider.js
@@ -152,7 +152,9 @@ class ConfigProvider extends EventEmitter {
         'homeEndBehaviour': true, // If checked (true), CodeMirror goes to start/end of a paragraph, not a line.
         'enableTableHelper': true, // Enable the table helper plugin
         'indentUnit': 4, // The number of spaces to be added
-        'countChars': false // Set to true to enable counting characters instead of words
+        'countChars': false, // Set to true to enable counting characters instead of words
+        'boldFormatting': '**', // Can be ** or __
+        'italicFormatting': '_' // Can be * or _
       },
       'display': {
         'theme': 'berlin', // The theme used by the app, can be berlin, frankfurt, bielefeld